+++ Changelog.md
@@ -1,7 +1,10 @@
 # Change Log
 
-## 1.0.0 - _26.10.2025_
+## 1.0.0 - _03.11.2025_
 
 **Release**
 
-- Initial release
+- Added a popup to set port other than 3210
+- Auto-reconnect on port changed (via save button)
+- Auto-retry in 5 sec. if the server doesn't connect
+- Various useful message log, error, warn, alert.